// src/interfaces/Test.ts
import { Document } from 'mongoose';

export interface IQuestion {
  id: string;
  question: string;
  options: string[];
  correctAnswer: string;
  explanation?: string;
}

export interface ITest extends Document {
  title: string;
  description: string;
  questions: IQuestion[];
  totalQuestions: number;
  timeLimit?: number; // in minutes
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}